+++ CHANGELOG.md
@@ -15,6 +15,24 @@ This changelog catalogs all notable changes made to the project. The format
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). Releases are 
 versioned in accordance with [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- Config files can be outputted in Kubernetes ConfigMap format (`servo config -f configmap`).
+- All Connectors can now dispatch events (previously only available to the `Servo` class).
+- `Optimizer` now includes an `api_url` property.
+- Event accessor `Connector.events` and `Connector.get_event`.
+
+### Removed
+- Removed `ServoAssembly.default_routes` from the API (usage eliminated).
+
+### Changed
+- Normalized naming of miscellaneous config methods.
+- Renamed `ServoAssembly.all_connectors` to `ServoAssembly.all_connector_types` for clarity.
+
+### Fixed
+- Pydantic will no longer see self-references between connectors (avoids recursion sharp edge).
+
 ## [0.2.5] - 2020-07-13
 
 ### Fixed